Separated userhome into different components
