+++ server/routes/matches.js
@@ -9,7 +9,11 @@ const db = pool
 */
 router.get("/users/:id", (req, res)=>{
     const user_id = req.params.id;
-    const getMatches = `SELECT Pets.name, Pets.pet_id FROM tinder_paws.Pets JOIN tinder_paws.Matches on Pets.pet_id = Matches.pet_id JOIN tinder_paws.Users on Users.user_id = Matches.user_id WHERE Users.user_id = ${user_id};`
+    const getMatches = `SELECT Pets.name, Pets.pet_id 
+    FROM tinder_paws.Pets 
+    JOIN tinder_paws.Matches on Pets.pet_id = Matches.pet_id 
+    JOIN tinder_paws.Users on Users.user_id = Matches.user_id 
+    WHERE Users.user_id = ${user_id};`
     db.query(getMatches, (err, result)=>{
         if(err){
             console.error(err.message);
